+++ lib/payloads/zeroConfUpdatePayload.js
@@ -1,4 +1,4 @@
-const { encryptationData } = require('../ewelink-helper');
+const { encryptationData } = require('../../src/helpers/ewelink');
 const { timestamp } = require('../helpers');
 
 const zeroConfUpdatePayload = (selfApikey, deviceId, deviceKey, params) => {
